+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 * No blinking on analysis pages if more than 1000 travels
 * Correctly handle not supported purposes
+* Correctly handle not supported manual modes
+
 
 ## 🔧 Tech